--- CryptoPWMS/Models/PasswordGroup.cs
@@ -1,8 +1,0 @@
-﻿namespace CryptoPWMS.Models
-{
-    public class PasswordGroup
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-}
