--- CryptoPWMS/Models/User.cs
@@ -1,9 +1,0 @@
-﻿namespace CryptoPWMS.Models
-{
-    public class User
-    {
-        public int Id { get; set; }
-        public string Username { get; set; }
-        public string Password { get; set; }
-    }
-}
